requirements.txt
gradio
torch
torchvision
ultralytics
opencv-python-headless
Pillow
numpy